# internal/services/identity_service/docs/v1/swagger.yaml
definitions:
...
  CreateOrEditUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
...
  CreateUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
...
  EditUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
...
  GetUserByIdResult:
    properties:
      user:
        $ref: '#/definitions/CreateOrEditUserDto'
    type: object
  GetUsersResult:
    properties:
      items:
        items:
          $ref: '#/definitions/UserDto'
        type: array
      totalCount:
        type: integer
    type: object
...
  UpdateUserPermissionDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
    type: object
  UserDto:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      userName:
        type: string
    type: object
...
  UserPermissionsResult:
    properties:
      items:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
...
  /api/v1/identities/user:
    post:
      consumes:
      - application/json
      description: Create new user
      operationId: CreateUser
      parameters:
      - description: CreateUserDto
        in: body
        name: CreateUserDto
        schema:
          $ref: '#/definitions/CreateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      operationId: UpdateUser
      parameters:
      - description: EditUserDto
        in: body
        name: EditUserDto
        schema:
          $ref: '#/definitions/EditUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
  /api/v1/identities/user/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete user
      operationId: DeleteUser
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user by id
      operationId: GetUserById
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetUserByIdResult'
      security:
      - ApiKeyAuth: []
      summary: Get user by id
      tags:
      - Users
  /api/v1/identities/user/{userId}/permissions:
    get:
      consumes:
      - application/json
      description: Get user permissions
      operationId: GetUserPermissions
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserPermissionsResult'
      security:
      - ApiKeyAuth: []
      summary: Get user permissions
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user permissions
      operationId: UpdateUserPermissions
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      - description: UpdateUserPermissionDto
        in: body
        name: UpdateUserPermissionDto
        schema:
          $ref: '#/definitions/UpdateUserPermissionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update user permissions
      tags:
      - Users
  /api/v1/identities/user/{userId}/reset-permissions:
    put:
      consumes:
      - application/json
      description: Reset user permissions
      operationId: ResetUserPermissions
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Reset user permissions
      tags:
      - Users
  /api/v1/identities/users:
    get:
      consumes:
      - application/json
      description: Get all users
      operationId: GetAllUsers
      parameters:
      - in: query
        name: filters
        type: string
      - in: query
        name: maxResultCount
        type: integer
      - in: query
        name: skipCount
        type: integer
      - in: query
        name: sorting
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetUsersResult'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"