# internal/services/identity_service/users/features/updating_user/v1/endpoints/updating_user_endpoint.go
package endpoints

import (
	"net/http"

	"github.com/go-playground/validator/v10"
	"github.com/jinzhu/copier"
	"github.com/labstack/echo/v4"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/events"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/http/echo/middlewares"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/permissions"
	postgresGorm "github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/postgres_gorm"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/pkg/rabbitmq"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/services/identity_service/users/dtos"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/services/identity_service/users/models"
	"github.com/tguankheng016/go-ecommerce-microservice/internal/services/identity_service/users/services"
)

func MapRoute(echo *echo.Echo, validator *validator.Validate, permissionManager services.IUserRolePermissionManager, rabbitMQPublisher rabbitmq.IPublisher) {
	group := echo.Group("/api/v1/identities/user")
	group.PUT("", updateUser(validator, permissionManager, rabbitMQPublisher), middlewares.Authorize(permissions.PagesAdministrationUsersEdit))
}

// @ID UpdateUser
// @Tags Users
// @Summary Update user
// @Description Update user
// @Accept json
// @Produce json
// @Param EditUserDto body EditUserDto false "EditUserDto"
// @Success 200 {object} UserDto
// @Security ApiKeyAuth
// @Router /api/v1/identities/user [put]
func updateUser(validator *validator.Validate, permissionManager services.IUserRolePermissionManager, rabbitMQPublisher rabbitmq.IPublisher) echo.HandlerFunc {
	return func(c echo.Context) error {
		ctx := c.Request().Context()

		var editUserDto dtos.EditUserDto

		if err := c.Bind(&editUserDto); err != nil {
			return c.JSON(http.StatusBadRequest, err)
		}

		if err := validator.StructCtx(ctx, editUserDto); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		tx, err := postgresGorm.GetTxFromContext(c)
		if err != nil {
			return echo.NewHTTPError(http.StatusInternalServerError, err)
		}

		userManager := services.NewUserManager(tx)

		var user models.User
		if err := tx.First(&user, editUserDto.Id).Error; err != nil {
			return echo.NewHTTPError(http.StatusNotFound, err)
		}

		if err := copier.Copy(&user, &editUserDto); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		if err := userManager.UpdateUser(&user, editUserDto.Password); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		if err := userManager.UpdateUserRoles(&user, editUserDto.RoleIds); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		permissionManager.RemoveUserRoleCaches(ctx, user.Id)

		var userDto dtos.UserDto
		if err := copier.Copy(&userDto, &user); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		var userUpdatedEvent events.UserUpdatedEvent
		if err := copier.Copy(&userUpdatedEvent, &user); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}
		if err := rabbitMQPublisher.PublishMessage(&userUpdatedEvent); err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, err)
		}

		return c.JSON(http.StatusOK, userDto)
	}
}